move layer tree code to specific folder
