+++ kadas/app/bullseye/kadasmaptoolbullseye.cpp
@@ -24,8 +24,8 @@
 #include <qgis/qgsunittypes.h>
 
 #include <kadas/gui/kadaslayerselectionwidget.h>
-#include <kadas/app/bullseye/kadasbullseyelayer.h>
-#include <kadas/app/bullseye/kadasmaptoolbullseye.h>
+#include <bullseye/kadasbullseyelayer.h>
+#include <bullseye/kadasmaptoolbullseye.h>
 
 KadasMapToolBullseye::KadasMapToolBullseye( QgsMapCanvas *canvas, QgsLayerTreeView *layerTreeView, QgsMapLayer *layer )
   : QgsMapTool( canvas )
