+++ kadas/app/globe/featuresource/kadasglobefeaturesource.cpp
@@ -26,8 +26,8 @@
 
 #include <kadas/gui/kadasitemlayer.h>
 #include <kadas/gui/mapitems/kadasgeometryitem.h>
-#include <kadas/app/globe/featuresource/kadasglobefeaturesource.h>
-#include <kadas/app/globe/featuresource/kadasglobefeatureutils.h>
+#include <globe/featuresource/kadasglobefeaturesource.h>
+#include <globe/featuresource/kadasglobefeatureutils.h>
 
 
 osgEarth::Status KadasGlobeFeatureSource::initialize( const osgDB::Options *dbOptions )
