+++ kadas/app/globe/kadasglobebillboardmanager.cpp
@@ -21,7 +21,7 @@
 
 #include <kadas/gui/kadasmapcanvasitemmanager.h>
 #include <kadas/gui/mapitems/kadasmapitem.h>
-#include <kadas/app/globe/kadasglobebillboardmanager.h>
+#include <globe/kadasglobebillboardmanager.h>
 
 
 KadasGlobeBillboardManager::KadasGlobeBillboardManager( QObject *parent )
