+++ kadas/app/globe/kadasglobevectorlayerproperties.cpp
@@ -19,7 +19,7 @@
 
 #include <kadas/gui/kadasitemlayer.h>
 #include <kadas/gui/milx/kadasmilxlayer.h>
-#include <kadas/app/globe/kadasglobevectorlayerproperties.h>
+#include <globe/kadasglobevectorlayerproperties.h>
 
 
 Q_DECLARE_METATYPE( KadasGlobeVectorLayerConfig * )
