+++ kadas/app/kadasmessagelogviewer.cpp
@@ -14,7 +14,7 @@
  *                                                                         *
  ***************************************************************************/
 
-#include <kadas/app/kadasmessagelogviewer.h>
+#include <kadasmessagelogviewer.h>
 
 KadasMessageLogViewer::KadasMessageLogViewer( QWidget *parent, Qt::WindowFlags fl )
   : QgsMessageLogViewer( parent, fl )
