+++ kadas/app/kml/kadaskmlexport.cpp
@@ -30,8 +30,8 @@
 #include <qgis/qgsunittypes.h>
 
 #include <kadas/gui/kadasitemlayer.h>
-#include <kadas/app/kml/kadaskmlexport.h>
-#include <kadas/app/kml/kadaskmllabeling.h>
+#include <kml/kadaskmlexport.h>
+#include <kml/kadaskmllabeling.h>
 
 bool KadasKMLExport::exportToFile( const QString &filename, const QList<QgsMapLayer *> &layers, double exportScale, const QgsCoordinateReferenceSystem &mapCrs, const QgsRectangle &exportMapRect )
 {
