+++ kadas/app/layertree/kadaslayertreeviewtemporalindicator.cpp
