+++ kadas/app/layertree/kadaslayertreeviewtemporalindicator.h
